完善chunk的自动加载

+++ WorldManager/Location.h
@@ -7,9 +7,9 @@ class location
 private:
 	world* _world;
 
-	int32_t _x;
-	int32_t _y;
-	int32_t _z;
+	double _x;
+	double _y;
+	double _z;
 
 public:
 	world* getWorld() const
@@ -17,32 +17,32 @@ public:
 		return _world;
 	}
 
-	int32_t getX() const
+	double getX() const
 	{
 		return _x;
 	}
-	int32_t getY() const
+	double getY() const
 	{
 		return _y;
 	}
-	int32_t getZ() const
+	double getZ() const
 	{
 		return _z;
 	}
 
 	int32_t getChunkX() const
 	{
-		return _x > 0 ? _x / 16 : _x / 16 - 1;
+		return _x > 0 ? (int32_t)_x / 16 : (int32_t)_x / 16 - 1;
 	}
 	int32_t getChunkY() const
 	{
-		return _y / 16;
+		return (int32_t)_y / 16;
 	}
 	int32_t getChunkZ() const
 	{
-		return _z > 0 ? _z / 16 : _z / 16 - 1;
+		return _z > 0 ? (int32_t)_z / 16 : (int32_t)_z / 16 - 1;
 	}
-	location(world* world, int32_t x, int32_t y, int32_t z)
+	location(world* world, double x, double y, double z)
 	{
 		_world = world;
